Cleaned up the project and tried to make progress with various problems.
Cleaned up the project so individual codes are within their own
associated directories. Also made progress within some of the codes.
Brachistochrone problem is still not properly tackled. Seems as though I
need to impose a continuity constraint somehow if I want to reproduce
the classical solution. The numeric solution I get currenly makes sense
as optimally minimizing the cost function I have defined (which is
essentially the minimum time integral), but it isn't smooth and thus
does not match that of the classical cycloid solution. Will need to
investigate.